Allow separate compilation
