+++ ModelicaCompliance/Functions/ExternalObjects/ExternalObjectTable.mo
@@ -11,14 +11,14 @@ model ExternalObjectTable
       output MyTable outTable;
 
       external "C" outTable = initMyTable(vals, size(vals, 1));
-      annotation(Include="#include \"ExtObj.c\"");
+      annotation(Include="#include \"ExtObjInit.c\"");
     end constructor;
 
     function destructor
       input MyTable inTable;
 
       external "C" closeMyTable(inTable) ;
-      annotation(Include="#include \"ExtObj.c\"");
+      annotation(Include="#include \"ExtObjClose.c\"");
     end destructor;
   end MyTable;
 
@@ -28,7 +28,7 @@ model ExternalObjectTable
     output Real y;
 
     external "C" y=interpolateMyTable(interpolTable,u) ;
-    annotation(Include="#include \"ExtObj.c\"");
+    annotation(Include="#include \"ExtObjInterpolate.c\"");
   end interpolateMyTable;
 
   MyTable myTable = MyTable({5, 3, 1, 9, 6});
